+++ src/GeneratorInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Fluwork\CodeGenerator;
+
+interface GeneratorInterface
+{
+    public function generate(): string;
+}