+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 This changelog follows [**Keep a changelog** conventions](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+
+### Added
+
+- `FileGenerator` used to generate PHP files